Consolidate L4 Match generator utils
This commit moves common functionality from
netpol and ANP into utils so that moving
forward the utils are all shared across.
