+++ go-controller/pkg/libovsdb/util/acl.go
@@ -255,3 +255,70 @@ func GetNetworkPolicyPort(proto v1.Protocol, port, endPort int32) *NetworkPolicy
 		EndPort:  endPort,
 	}
 }
+
+// gressRulePortsForL4ACLMatch for a given ingress/egress rule in a policy,
+// captures all the provided port ranges and individual ports
+// so that using this, ACL L4 matches can be generated
+type gressRulePortsForL4ACLMatch struct {
+	portList  []string // list of provided ports as string
+	portRange []string // list of provided port ranges in OVN ACL format
+}
+
+func getProtocolPortsMap(rulePorts []*NetworkPolicyPort) map[string]*gressRulePortsForL4ACLMatch {
+	gressProtoPortsMap := make(map[string]*gressRulePortsForL4ACLMatch)
+	for _, pp := range rulePorts {
+		gpp, ok := gressProtoPortsMap[pp.Protocol]
+		if !ok {
+			gpp = &gressRulePortsForL4ACLMatch{portList: []string{}, portRange: []string{}}
+			gressProtoPortsMap[pp.Protocol] = gpp
+		}
+		if pp.EndPort != 0 && pp.EndPort != pp.Port {
+			gpp.portRange = append(gpp.portRange, fmt.Sprintf("%d<=%s.dst<=%d", pp.Port, pp.Protocol, pp.EndPort))
+		} else if pp.Port != 0 {
+			gpp.portList = append(gpp.portList, fmt.Sprintf("%d", pp.Port))
+		}
+	}
+	return gressProtoPortsMap
+}
+
+func getL4Match(protocol string, ports *gressRulePortsForL4ACLMatch) string {
+	allL4Matches := []string{}
+	if len(ports.portList) > 0 {
+		// if there is just one port, then don't use `{}`
+		template := "%s.dst==%s"
+		if len(ports.portList) > 1 {
+			template = "%s.dst=={%s}"
+		}
+		allL4Matches = append(allL4Matches, fmt.Sprintf(template, protocol, strings.Join(ports.portList, ",")))
+	}
+	allL4Matches = append(allL4Matches, ports.portRange...)
+	l4Match := protocol
+	if len(allL4Matches) > 0 {
+		template := "%s && %s"
+		if len(allL4Matches) > 1 {
+			template = "%s && (%s)"
+		}
+		l4Match = fmt.Sprintf(template, protocol, strings.Join(allL4Matches, " || "))
+	}
+	return l4Match
+}
+
+// GetL4MatchesFromNetworkPolicyPorts accepts a list of NetworkPolicyPorts cache and
+// constructs l4Matches for each protocol type
+// It returns a map that has protocol as the key and the l4Match as the value
+// If len(rulePorts)==0; it returns map["None"] = "None" which means there is no L4 match
+func GetL4MatchesFromNetworkPolicyPorts(rulePorts []*NetworkPolicyPort) map[string]string {
+	l4Matches := make(map[string]string)
+	gressProtoPortsMap := getProtocolPortsMap(rulePorts)
+	if len(gressProtoPortsMap) == 0 {
+		gressProtoPortsMap[UnspecifiedL4Protocol] = nil
+	}
+	for protocol, ports := range gressProtoPortsMap {
+		l4Match := UnspecifiedL4Match
+		if ports != nil {
+			l4Match = getL4Match(protocol, ports)
+		}
+		l4Matches[protocol] = l4Match
+	}
+	return l4Matches
+}
